backend/requirements.txt
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dill==0.3.7
docopt==0.6.2
Flask==3.0.0
Flask-Cors==4.0.0
fonttools==4.44.0
google-api-core==2.14.0
google-auth==2.23.4
google-cloud-aiplatform==1.36.3
google-cloud-bigquery==3.13.0
google-cloud-core==2.3.3
google-cloud-resource-manager==1.10.4
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpc-google-iam-v1==0.12.7
grpcio==1.59.2
grpcio-status==1.59.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lightgbm==4.1.0
llvmlite==0.41.1
MarkupSafe==2.1.3
matplotlib==3.8.1
numba==0.58.1
numpy==1.26.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
pipreqs==0.4.13
plotly==5.18.0
proto-plus==1.22.3
protobuf==4.25.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
shap==0.43.0
shapely==2.0.2
six==1.16.0
slicer==0.0.7
tenacity==8.2.3
threadpoolctl==3.2.0
tqdm==4.66.1
tzdata==2023.3
urllib3==2.0.7
vertexai==0.0.1
Werkzeug==3.0.1
yarg==0.1.9
